features: Implemented BouquetController methods, implemented BouquetService methods.

+++ server/src/main/java/com/example/flowersproject/repository/BouquetRepository.java
@@ -4,6 +4,6 @@ import com.example.flowersproject.entity.products.BouquetEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface BouquetRepository extends JpaRepository<BouquetEntity, Integer> {
+public interface BouquetRepository extends JpaRepository<BouquetEntity, Long> {
 
 }
